Issue #132 Cosmetic tweaks for console output in parler_po commands
